Splitte opp selve headeren og menypunktene til forskjellige komponenter for arbeidsgiver og personbruker
