+++ pam-frontend-header/src/arbeidsgiver/ArbeidsgiverSelect.less
@@ -1,0 +1,9 @@
+@import "../../node_modules/nav-frontend-core/less/_variabler";
+
+.ArbeidsgiverSelect {
+    margin: 0 !important;
+    width: 100%;
+    label {
+        margin: 0;
+    }
+}
