--- pam-frontend-header/src/arbeidsgiver/HeaderMeny.less
@@ -1,98 +1,0 @@
-@import "../../node_modules/nav-frontend-core/less/_variabler";
-@import "../styles.less";
-
-.ArbeidsgiverHeaderMeny__wrapper {
-  background-color: white;
-  padding-left: 2rem;
-  padding-right: 2rem;
-  border-bottom: 2px solid @yellowLight;
-}
-
-.ArbeidsgiverHeaderMeny {
-  color: #4a4a4a;
-  font-family: "Source Sans Pro", Arial, sans-serif;
-  display: flex;
-  flex-direction: column;
-  max-width: 1156px;
-  margin: auto;
-  padding: 0 8px;
-
-  .topp {
-    padding: 2.5rem 0 2rem 0;
-    display: flex;
-    align-items: center;
-
-    > :first-child {
-      flex: 1;
-    }
-
-    > :not(:first-child) {
-      margin-left: 1rem;
-    }
-
-    .logo {
-      font-size: 24px;
-      font-weight: 600;
-      a {
-        color: #4a4a4a;
-        text-decoration: none;
-      }
-    }
-
-    .topmeny-select {
-      margin: 0 !important;
-      width: 100%;
-      label {
-        margin: 0
-      }
-    }
-  }
-
-  .meny {
-    display: flex;
-    align-items: start;
-    justify-content: center;
-
-    .meny--lenke:focus {
-      outline: none;
-      > li {
-        > div {
-          background-color: @orangeFocus;
-        }
-      }
-    }
-
-    ul {
-      display: flex;
-      margin: 0;
-    }
-
-    li {
-      list-style: none;
-      margin: 0 2rem;
-      font-size: 16px;
-      padding: 4px 2px 22px 2px;
-      text-transform: uppercase;
-      font-weight: bold;
-      color: @darkBlue;
-      &.active {
-        border-bottom: 8px solid @yellowLight;
-        color: @darkBlue;
-        a {
-          &:hover {
-            text-decoration: none;
-          }
-        }
-      }
-      &.not-active {
-        &:hover {
-          border-bottom: 8px solid @yellowLight;
-        }
-      }
-    }
-
-    a {
-      text-decoration: none;
-    }
-  }
-}
