+++ pam-frontend-header/src/arbeidsgiver/VelgArbeidsgiver.less
@@ -1,6 +1,11 @@
 .container-arbeidsgiver {
   padding-top: 2.5rem;
   max-width: 700px;
+
+  .ArbeidsgiverSelect {
+    width: 210px;
+    margin: 0 auto !important;
+  }
 }
 
 .panel.panel--arbeidsgiver {
