+++ pam-frontend-header/src/felles/popover/Popover.less
