+++ pam-frontend-header/src/index.tsx
@@ -1,24 +1,20 @@
-import { ArbeidsgiverHeaderMeny, ArbeidsgiverTabId } from './arbeidsgiver/HeaderMeny';
-import { ArbeidsgiverHeader } from './arbeidsgiver/Header';
 import { VelgArbeidsgiver } from './arbeidsgiver/VelgArbeidsgiver';
 import { VeilederHeaderMeny, VeilederTabId } from './veileder/HeaderMeny';
-import { PersonbrukerHeaderMeny } from './personbruker/PersonbrukerHeaderMeny';
-import { PersonbrukerHeader } from './personbruker/PersonbrukerHeader';
-import { PersonbrukerApplikasjon } from './personbruker/InnloggetMeny';
-import { HeaderUtenMenypunkter } from './personbruker/HeaderUtenMenypunkter';
-import { PersonbrukerHeaderMedLoggUt } from './personbruker/PersonbrukerHeaderMedLoggUt';
+import { Header, AuthStatus } from './felles/Header';
+import { Personbrukermeny, PersonbrukerApplikasjon } from './personbruker/Personbrukermeny';
+import { Arbeidsgivermeny, ArbeidsgiverTabId } from './arbeidsgiver/Arbeidsgivermeny';
+import ArbeidsgiverSelect from './arbeidsgiver/ArbeidsgiverSelect';
 
 export {
-    ArbeidsgiverHeader,
-    ArbeidsgiverHeaderMeny,
     ArbeidsgiverTabId,
     VelgArbeidsgiver,
+    ArbeidsgiverSelect,
     VeilederHeaderMeny,
     VeilederTabId,
-    PersonbrukerHeaderMeny,
-    PersonbrukerHeader,
-    HeaderUtenMenypunkter,
-    PersonbrukerHeaderMedLoggUt,
-    PersonbrukerApplikasjon
+    PersonbrukerApplikasjon,
+    Header,
+    AuthStatus,
+    Personbrukermeny,
+    Arbeidsgivermeny
 }
