--- pam-frontend-header/src/personbruker/PersonbrukerHeader.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import './PersonbrukerHeaderMeny.less';
-
-export const PersonbrukerHeader = () => (
-  <div className="PersonbrukerHeaderMeny">
-    <div className="topp">
-      <div className="logo">
-        <a href="/">Arbeidsplassen</a>
-      </div>
-    </div>
-  </div>
-);
