+++ pam-frontend-header/src/personbruker/Personbrukermeny.less
@@ -1,0 +1,106 @@
+@import "../../node_modules/nav-frontend-core/less/_variabler";
+@import "../variables.less";
+
+.Personbrukermeny {
+	padding-bottom: 0;
+
+	a {
+		text-decoration: none;
+		border-bottom: none;
+	}
+}
+
+.Personbrukermeny--lenke {
+  padding: 1rem;
+  display: block;
+  background-color: transparent;
+}
+
+.Personbrukermeny--lenke-inner {
+	display: flex;
+	justify-content: space-between;
+	padding-bottom: 0;
+	text-transform: uppercase;
+	font-weight: bold;
+	color: @darkBlue;
+}
+
+.Personbrukermeny--lenke-active {
+	background-color: @yellowLight2;
+}
+
+.Personbrukermeny--lenke-wrapper, .Personbrukermeny--lenke-wrapper-CV {
+	border-bottom: 1px solid #dbdbdb;
+}
+
+@media(min-width: @screen-sm-min) {
+	.Personbrukermeny {
+		display: flex;
+		justify-content: space-between;
+		max-width: 800px;
+		margin: auto;
+	}
+
+	.Personbrukermeny--lenke-inner,
+	.Personbrukermeny--navn-inner {
+		&:focus {
+				outline: none;
+				background-color: transparent;
+		}
+	}
+
+	.Personbrukermeny--lenke:focus .Personbrukermeny--lenke,
+	.Personbrukermeny--navn:focus .Personbrukermeny--navn {
+		background-color: @orangeFocus; /* keyboard-only focus styles */
+		outline: none;
+	}
+
+	.Personbrukermeny--lenke {
+		font-size: 16px;
+		text-transform: uppercase;
+		padding: 0;
+		color: @navMorkGra;
+	}
+
+	.Personbrukermeny--lenke-inner {
+		padding: 8px 2px 22px 2px;
+		border-bottom: 8px solid transparent;
+		font-weight: bold;
+		color: @darkBlue;
+	}
+
+	.Personbrukermeny--lenke:hover .Personbrukermeny--lenke-inner {
+		border-bottom: 8px solid @yellowLight;
+	}
+
+	.Personbrukermeny--lenke-active .Personbrukermeny--lenke-inner,
+	.Personbrukermeny--lenke:active .Personbrukermeny--lenke-inner {
+		border-bottom: 8px solid @yellowLight;
+		font-weight: bold;
+		color: @darkBlue;
+		background-color: #fff;
+	}
+
+	.Personbrukermeny--lenke-active:hover .Personbrukermeny--lenke-inner {
+		border-bottom: 8px solid @yellowLight;
+		color: @darkBlue;
+	}
+
+	.Personbrukermeny--lenke-wrapper {
+		border-bottom: none;
+	}
+
+  .Personbrukermeny--lenke-wrapper-CV {
+		border-bottom: none;
+		margin: 0 -15px;
+
+		.Personbrukermeny--lenke-inner {
+			padding-left: 15px;
+			padding-right: 15px;
+		}
+	}
+
+	.Personbrukermeny__Innstillinger--mobile {
+		display: none;
+	}
+}
